Use custom Babel loader to avoid using separate Babel copies for loader and loader options (#4417)
This resolves the

> .value is not a valid Plugin property

error showing up for people in https://github.com/zeit/next.js/issues/4227

cc @timneutkens

+++ test/integration/basic/test/hmr.js
@@ -1,6 +1,6 @@
 /* global describe, it, expect */
 import webdriver from 'next-webdriver'
-import { readFileSync, writeFileSync, renameSync } from 'fs'
+import { readFileSync, writeFileSync, renameSync, existsSync } from 'fs'
 import { join } from 'path'
 import { waitFor, check } from 'next-test-utils'
 import cheerio from 'cheerio'
@@ -9,33 +9,39 @@ export default (context, renderViaHTTP) => {
   describe('Hot Module Reloading', () => {
     describe('delete a page and add it back', () => {
       it('should load the page properly', async () => {
-        const browser = await webdriver(context.appPort, '/hmr/contact')
-        const text = await browser
-          .elementByCss('p').text()
-        expect(text).toBe('This is the contact page.')
-
         const contactPagePath = join(__dirname, '../', 'pages', 'hmr', 'contact.js')
         const newContactPagePath = join(__dirname, '../', 'pages', 'hmr', '_contact.js')
 
-        // Rename the file to mimic a deleted page
-        renameSync(contactPagePath, newContactPagePath)
+        try {
+          const browser = await webdriver(context.appPort, '/hmr/contact')
+          const text = await browser
+            .elementByCss('p').text()
+          expect(text).toBe('This is the contact page.')
 
-        // wait until the 404 page comes
-        await check(
-          () => browser.elementByCss('body').text(),
-          /This page could not be found/
-        )
+          // Rename the file to mimic a deleted page
+          renameSync(contactPagePath, newContactPagePath)
 
-        // Rename the file back to the original filename
-        renameSync(newContactPagePath, contactPagePath)
+          // wait until the 404 page comes
+          await check(
+            () => browser.elementByCss('body').text(),
+            /(This page could not be found|ENOENT)/
+          )
 
-        // wait until the page comes back
-        await check(
-          () => browser.elementByCss('body').text(),
-          /This is the contact page/
-        )
+          // Rename the file back to the original filename
+          renameSync(newContactPagePath, contactPagePath)
 
-        browser.close()
+          // wait until the page comes back
+          await check(
+            () => browser.elementByCss('body').text(),
+            /This is the contact page/
+          )
+
+          browser.close()
+        } finally {
+          if (existsSync(newContactPagePath)) {
+            renameSync(newContactPagePath, contactPagePath)
+          }
+        }
       })
     })
